requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
colorlog==6.7.0
ez-setup==0.9
frozenlist==1.3.1
hikari==2.0.0.dev111
hikari-lightbulb==2.2.4
idna==3.4
multidict==6.0.2
python-dotenv==0.21.0
typing_extensions==4.3.0
uvloop==0.17.0
yarl==1.8.1
